[CCAP-685] Review cleanup
